Split testimage.py out of testraster.py
