+++ tests/testraster.py
@@ -14,18 +14,9 @@ class TestRaster(ComparisonTestCase):
     def test_raster_init(self):
         Raster(self.array1)
 
-    def test_image_init(self):
-        image = Image(self.array1)
-        self.assertEqual(image.xdensity, 3)
-        self.assertEqual(image.ydensity, 2)
-
     def test_raster_index(self):
         raster = Raster(self.array1)
         self.assertEqual(raster[0, 1], 3)
-
-    def test_image_index(self):
-        image = Image(self.array1)
-        self.assertEqual(image[-.33, -0.25], 3)
 
     def test_raster_sample(self):
         raster = Raster(self.array1)
@@ -33,29 +24,11 @@ class TestRaster(ComparisonTestCase):
                          Curve(np.array([(0, 0), (1, 1), (2, 2)]),
                                kdims=['x'], vdims=['z']))
 
-    def test_image_sample(self):
-        image = Image(self.array1)
-        self.assertEqual(image.sample(y=0.25),
-                         Curve(np.array([(-0.333333, 0), (0, 1), (0.333333, 2)]),
-                               kdims=['x'], vdims=['z']))
-
     def test_raster_range_masked(self):
         arr = np.random.rand(10,10)-0.5
         arr = np.ma.masked_where(arr<=0, arr)
         rrange = Raster(arr).range(2)
         self.assertEqual(rrange, (np.min(arr), np.max(arr)))
-
-    def test_image_range_masked(self):
-        arr = np.random.rand(10,10)-0.5
-        arr = np.ma.masked_where(arr<=0, arr)
-        rrange = Image(arr).range(2)
-        self.assertEqual(rrange, (np.min(arr), np.max(arr)))
-
-    def test_empty_image(self):
-        Image([])
-        Image(None)
-        Image(np.array([]))
-        Image(np.zeros((0, 0)))
 
 
 
@@ -74,7 +47,7 @@ class TestQuadMesh(ComparisonTestCase):
         self.assertEqual(qmesh.vdims, img.vdims)
         self.assertEqual(qmesh.group, img.group)
         self.assertEqual(qmesh.label, img.label)
-        
+
     def test_quadmesh_to_trimesh(self):
         qmesh = QuadMesh(([0, 1], [0, 1], np.array([[0, 1], [2, 3]])))
         trimesh = qmesh.trimesh()
